Bug that was affecting !src/!say fixed.
Marko had a glaring speech generation bug that was causing him to not distinguish between two words with the same id in a
sentence, when he pulled for the next word in a source sentence. This would cause looping and text to be less random /
bugged, which in turn affected !src's painting of colors. Both seem to be resolved now except for a src not keeping the
initial chain id.