Optimization for speech generation (random chain selection).
Random chain selection previously queried twice. Once to get a list of
list of tids and ids. Then again to get the count of the amount of ^
returned. Effectively, for large queries, the time was doubled.

Now the query is stored in a table, in memory, then counted and
shifted. There's a solid 1-2 second difference per query for well known
words.